+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.24] - 2024-09-14
+
+### Added
+
+- Test for valid positions for rook
+- Method to see if to_row and to-col are in valid_positions
+
+### Changed
+
+- Update valid_positions method for rook adding the horizontal moves
+
 ## [0.0.23] - 2024-09-13
 
 ### Added